refactor: simplified repo
